initial nodebalancer querys clean up
