+++ packages/manager/src/features/NodeBalancers/NodeBalancerDetail/NodeBalancerSummary/SummaryPanel.tsx
@@ -13,26 +13,18 @@ import * as React from 'react';
 
 import { Link } from 'src/components/Link';
 import { TagCell } from 'src/components/TagCell/TagCell';
-import {
-  useIsLkeEnterpriseEnabled,
-  useKubernetesBetaEndpoint,
-} from 'src/features/Kubernetes/kubeUtils';
+import { useKubernetesBetaEndpoint } from 'src/features/Kubernetes/kubeUtils';
 import { IPAddress } from 'src/features/Linodes/LinodesLanding/IPAddress';
 import { useIsResourceRestricted } from 'src/hooks/useIsResourceRestricted';
 import { useKubernetesClusterQuery } from 'src/queries/kubernetes';
 
 export const SummaryPanel = () => {
   const { isUsingBetaEndpoint } = useKubernetesBetaEndpoint();
-  const { isLkeEnterpriseLAFeatureEnabled } = useIsLkeEnterpriseEnabled();
 
   const { id } = useParams({
     from: '/nodebalancers/$id/summary',
   });
-  const { data: nodebalancer } = useNodeBalancerQuery(
-    Number(id),
-    Boolean(id),
-    isLkeEnterpriseLAFeatureEnabled
-  );
+  const { data: nodebalancer } = useNodeBalancerQuery(Number(id), Boolean(id));
   const { data: configs } = useAllNodeBalancerConfigsQuery(Number(id));
   const { data: regions } = useRegionsQuery();
   const { data: attachedFirewallData } = useNodeBalancersFirewallsQuery(
