+++ packages/queries/src/nodebalancers/index.ts
@@ -1,1 +1,2 @@
+export * from './keys';
 export * from './nodebalancers';
